+++ src/OpenDDD/Domain/Model/IRepository.cs
@@ -2,13 +2,15 @@
 
 namespace OpenDDD.Domain.Model
 {
-    public interface IRepository<TAggregateRoot, TId> where TAggregateRoot : IAggregateRoot<TId>
+    public interface IRepository<TAggregateRoot, in TId> 
+        where TAggregateRoot : IAggregateRoot<TId>
+        where TId : notnull
     {
-        Task<TAggregateRoot> GetAsync(TId id);
-        Task<TAggregateRoot?> FindAsync(TId id);
-        Task<IEnumerable<TAggregateRoot>> FindWithAsync(Expression<Func<TAggregateRoot, bool>> filterExpression);
-        Task<IEnumerable<TAggregateRoot>> FindAllAsync();
+        Task<TAggregateRoot> GetAsync(TId id, CancellationToken ct);
+        Task<TAggregateRoot?> FindAsync(TId id, CancellationToken ct);
+        Task<IEnumerable<TAggregateRoot>> FindWithAsync(Expression<Func<TAggregateRoot, bool>> filterExpression, CancellationToken ct);
+        Task<IEnumerable<TAggregateRoot>> FindAllAsync(CancellationToken ct);
         Task SaveAsync(TAggregateRoot aggregateRoot, CancellationToken ct);
-        Task DeleteAsync(TAggregateRoot aggregateRoot);
+        Task DeleteAsync(TAggregateRoot aggregateRoot, CancellationToken ct);
     }
 }