+++ src/OpenDDD/Main/Interfaces/IStartable.cs
@@ -1,0 +1,7 @@
+﻿namespace OpenDDD.Main.Interfaces
+{
+    public interface IStartable
+    {
+        Task StartAsync(CancellationToken cancellationToken);
+    }
+}