+++ src/OpenDDD/Main/Options/OpenDddPipelineOptions.cs
@@ -1,7 +1,0 @@
-﻿namespace OpenDDD.Main.Options
-{
-    public class OpenDddPipelineOptions
-    {
-        
-    }
-}