+++ src/OpenDDD/Main/Options/OpenDddServiceOptions.cs
@@ -1,8 +1,0 @@
-﻿namespace OpenDDD.Main.Options
-{
-    public class OpenDddServiceOptions
-    {
-        public bool AutoRegisterActions { get; set; } = true;
-        public bool AutoRegisterRepositories { get; set; } = true;
-    }
-}